Clean manual tests
